@@ -3,9 +3,10 @@ package config
 // ServiceProvider 实现文件 provider.go
 
 import (
+	"path/filepath"
+
 	"github.com/chenbihao/gob/framework"
 	"github.com/chenbihao/gob/framework/contract"
-	"path/filepath"
 )
 
 // ConfigProvider 服务提供者具体实现方法
@@ -14,14 +15,9 @@ type ConfigProvider struct {
 
 var _ framework.ServiceProvider = (*ConfigProvider)(nil)
 
-// Register 注册一个服务实例
-func (provider *ConfigProvider) Register(container framework.Container) framework.NewInstance {
-	return NewConfigService
-}
-
-// Boot 启动的时候注入
-func (provider *ConfigProvider) Boot(container framework.Container) error {
-	return nil
+// Name 定义对应的服务字符串凭证
+func (provider *ConfigProvider) Name() string {
+	return contract.ConfigKey
 }
 
 // IsDefer 是否延迟加载
@@ -29,18 +25,23 @@ func (provider *ConfigProvider) IsDefer() bool {
 	return false
 }
 
+// Boot 启动的时候注入
+func (provider *ConfigProvider) Boot(container framework.Container) error {
+	return nil
+}
+
 // Params 定义要传递给实例化方法的参数
 func (provider *ConfigProvider) Params(container framework.Container) []interface{} {
 	appService := container.MustMake(contract.AppKey).(contract.App)
 	envService := container.MustMake(contract.EnvKey).(contract.Env)
-	env := envService.AppEnv()
+	appEnv := envService.AppEnv()
 	// 配置文件夹地址
 	configFolder := appService.ConfigFolder()
-	envFolder := filepath.Join(configFolder, env)
-	return []interface{}{container, envFolder, envService.All()}
+	appEnvFolder := filepath.Join(configFolder, appEnv)
+	return []any{container, appEnvFolder, envService.All()}
 }
 
-// Name 定义对应的服务字符串凭证
-func (provider *ConfigProvider) Name() string {
-	return contract.ConfigKey
+// Register 注册一个服务实例
+func (provider *ConfigProvider) Register(container framework.Container) framework.NewInstance {
+	return NewConfigService
 }
