@@ -25,8 +25,8 @@ import (
 	"strings"
 	"time"
 
+	cobra "github.com/chenbihao/gob/framework/cobra"
 	"github.com/cpuguy83/go-md2man/v2/md2man"
-	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
 
@@ -133,7 +133,7 @@ func fillHeader(header *GenManHeader, name string, disableAutoGen bool) error {
 		}
 		header.Date = &now
 	}
-	header.date = (*header.Date).Format("Jan 2006")
+	header.date = header.Date.Format("Jan 2006")
 	if header.Source == "" && !disableAutoGen {
 		header.Source = "Auto generated by spf13/cobra"
 	}
