@@ -17,7 +17,7 @@ package doc
 import (
 	"strings"
 
-	"github.com/spf13/cobra"
+	"github.com/chenbihao/gob/framework/cobra"
 )
 
 // Test to see if we have a reason to print See Also information in docs
@@ -40,7 +40,7 @@ func hasSeeAlso(cmd *cobra.Command) bool {
 // that do not contain \n.
 func forceMultiLine(s string) string {
 	if len(s) > 60 && !strings.Contains(s, "\n") {
-		s = s + "\n"
+		s += "\n"
 	}
 	return s
 }
