@@ -14,14 +14,9 @@ type AppProvider struct {
 
 var _ framework.ServiceProvider = (*AppProvider)(nil)
 
-// Register 注册 AppService 方法
-func (provider *AppProvider) Register(container framework.Container) framework.NewInstance {
-	return NewGobApp
-}
-
-// Boot 启动调用
-func (provider *AppProvider) Boot(container framework.Container) error {
-	return nil
+// Name 获取字符串凭证
+func (provider *AppProvider) Name() string {
+	return contract.AppKey
 }
 
 // IsDefer 是否延迟初始化
@@ -29,12 +24,17 @@ func (provider *AppProvider) IsDefer() bool {
 	return false
 }
 
+// Boot 启动调用
+func (provider *AppProvider) Boot(container framework.Container) error {
+	return nil
+}
+
 // Params 获取初始化参数
 func (provider *AppProvider) Params(container framework.Container) []interface{} {
 	return []interface{}{container, provider.BaseFolder}
 }
 
-// Name 获取字符串凭证
-func (provider *AppProvider) Name() string {
-	return contract.AppKey
+// Register 注册 AppService 方法
+func (provider *AppProvider) Register(container framework.Container) framework.NewInstance {
+	return NewGobApp
 }
